requirements.txt
acres==0.2.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
Django==4.2.16
django-autoslug==1.9.9
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.4
django-bootstrap5==24.3
djangorestframework==3.15.2
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.16.1
frontend==0.0.3
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
isodate==0.7.2
itsdangerous==2.2.0
looseversion==1.3.0
lxml==5.3.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.9.2
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==10.3.0
pluggy==1.5.0
prov==2.0.1
puremagic==1.28
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
PyMuPDF==1.25.2
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
rdflib==6.2.0
reportlab==4.2.5
requests==2.32.3
scipy==1.13.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
starlette==0.45.2
tomli==2.2.1
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
